+++ hw/xbox/nv2a/pgraph/glsl/geom.c
@@ -19,7 +19,6 @@
  * License along with this library; if not, see <http://www.gnu.org/licenses/>.
  */
 
-#include "hw/xbox/nv2a/pgraph/shaders.h"
 #include "common.h"
 #include "geom.h"
