+++ hw/xbox/nv2a/pgraph/glsl/meson.build
@@ -2,6 +2,7 @@ specific_ss.add([files(
 	'common.c',
 	'geom.c',
 	'psh.c',
+	'shaders.c',
 	'vsh.c',
 	'vsh-ff.c',
 	'vsh-prog.c',
