+++ hw/xbox/nv2a/pgraph/meson.build
@@ -3,7 +3,6 @@ specific_ss.add(files(
 	'profile.c',
 	'rdi.c',
 	's3tc.c',
-	'shaders.c',
 	'swizzle.c',
 	'texture.c',
 	'vertex.c',
