+++ hw/xbox/nv2a/pgraph/psh.h
@@ -51,40 +51,4 @@ enum ConvolutionFilter {
     CONVOLUTION_FILTER_GAUSSIAN,
 };
 
-typedef struct PshState {
-    uint32_t combiner_control;
-    uint32_t shader_stage_program;
-    uint32_t other_stage_input;
-    uint32_t final_inputs_0;
-    uint32_t final_inputs_1;
-
-    uint32_t rgb_inputs[8], rgb_outputs[8];
-    uint32_t alpha_inputs[8], alpha_outputs[8];
-
-    bool point_sprite;
-    bool rect_tex[4];
-    bool snorm_tex[4];
-    bool compare_mode[4][4];
-    bool alphakill[4];
-    int colorkey_mode[4];
-    enum ConvolutionFilter conv_tex[4];
-    bool tex_x8y24[4];
-    int dim_tex[4];
-
-    float border_logical_size[4][3];
-    float border_inv_real_size[4][3];
-
-    bool shadow_map[4];
-    enum PshShadowDepthFunc shadow_depth_func;
-
-    bool alpha_test;
-    enum PshAlphaFunc alpha_func;
-
-    bool window_clip_exclusive;
-
-    bool smooth_shading;
-    bool depth_clipping;
-    bool z_perspective;
-} PshState;
-
 #endif
